Paramgen improvements and naive paramgen fixes

+++ naive-paramgen/paramgen-queries/pg-19a.sql
@@ -2,5 +2,5 @@ SELECT
     city1Id AS 'city1Id:ID',
     city2Id AS 'city2Id:ID'
 FROM cityPairsNumFriends
-ORDER BY md5(3532569367*city1Id + 342663089*city2Id)
+ORDER BY md5(3532569367::bigint*city1Id + 342663089::bigint*city2Id)
 LIMIT 400
